Got quite far, it works and has no seqmentation faults but currently doesn't move

+++ src/qbot/brains.hpp
@@ -8,7 +8,7 @@
 class Brains {
 
 public:
-  Brains(short int precision, b2World* world);
+  Brains(int precision, b2World* world);
   ~Brains();
   
   Wurm* GetWurm();
